+++ test/integration/test_cases/web_content/cookbook/rendering/random_gradient_pattern.mjs
@@ -13,7 +13,7 @@ const testSteps =
 			return x - Math.floor(x);
 		}
 
-		chart.on('background-draw', event => {
+		chart.on('draw-begin', event => {
 			seed = 1;
 		});